Relationship tables finished + auth fixed + excpetions fixed

+++ EntitiesAPI/spring-server/src/main/java/ch/heigvd/amt/project2/repositories/ICinemaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ICinemaRepository extends CrudRepository<CinemaEntity, Long> {
+public interface ICinemaRepository extends CrudRepository<CinemaEntity, Integer> {
 
     Page<CinemaEntity> findAll(Pageable pageable);
